Project_4/requirements.txt
fastapi==0.85.1
passlib==1.7.4
pydantic==1.10.2
SQLAlchemy==1.4.42
pymysql
python-decouple
psycopg2-binary
aiofiles
jinja2